% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{rbind_list_base}
\alias{rbind_list_base}
\title{Bind list of data frames together}
\usage{
rbind_list_base(x)
}
\arguments{
\item{x}{List of data.frames to bind together}
}
\description{
Bind list of data frames together
}
